AI/requirements.txt
pandas
numpy
scikit-learn
tensorflow
sentence-transformers
google-generativeai
joblib
torch
python-dotenv
os
asyncio
aiomysql
bcrypt
re
ast
random
datetime
pathlib
fastapi
uvicorn[standard]
pydantic
aiomysql
mysql-connector-python
PyJWT
npm